app/requirements.txt
fastapi
uvicorn
pydantic==2.8.2
numpy
pandas
scikit-learn
python-dateutil
tldextract
PyYAML
textdistance
